--- lib/src/course/presentation/bloc/main/main_event.dart
@@ -1,5 +1,0 @@
-sealed class MainEvent {}
-
-final class MainUnfocusEvent extends MainEvent {}
-
-final class MainTouchEvent extends MainEvent {}
