--- lib/src/course/presentation/bloc/map/map_state.dart
@@ -1,16 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:solo_play_application/src/course/domain/models/map_model.dart';
-
-final class MapState extends Equatable {
-  final MapModel mapModel;
-  const MapState({required this.mapModel});
-
-  @override
-  List<Object?> get props => [mapModel];
-
-  MapState copyWith({
-    required MapModel mapModel,
-  }) {
-    return MapState(mapModel: mapModel);
-  }
-}
