--- lib/src/course/presentation/page/main_page.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:solo_play_application/src/course/presentation/bloc/main/main_bloc.dart';
-import 'package:solo_play_application/src/course/presentation/view/main_view.dart';
-
-class MainPage extends StatelessWidget {
-  const MainPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider(
-      create: (context) => MainBloc(),
-      child: const MainView(),
-    );
-  }
-}
