--- lib/src/post/presentation/widget/post_widget.dart
@@ -1,126 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:solo_play_application/src/style/theme_color.dart';
-
-class PostWidget extends StatefulWidget {
-  const PostWidget({super.key});
-
-  @override
-  State<PostWidget> createState() => _PostWidgetState();
-}
-
-class _PostWidgetState extends State<PostWidget> {
-  final PageController _controller = PageController();
-  int _currentPage = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Container(
-          width: 175,
-          height: 164,
-          decoration: BoxDecoration(
-            borderRadius: const BorderRadius.vertical(
-              top: Radius.circular(10),
-              bottom: Radius.zero,
-            ),
-            color: ThemeColor.basicGrey,
-          ),
-          child: Stack(
-            children: [
-              ClipRRect(
-                borderRadius:
-                    const BorderRadius.vertical(top: Radius.circular(10)),
-                child: PageView(
-                  controller: _controller,
-                  onPageChanged: (index) {
-                    setState(() {
-                      _currentPage = index;
-                    });
-                  },
-                  children: [
-                    Container(color: Colors.red),
-                    Container(color: Colors.green),
-                    Container(color: Colors.blue),
-                  ],
-                ),
-              ),
-              Positioned(
-                bottom: 8.0,
-                right: 0,
-                left: 0,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: List.generate(
-                    3,
-                    (index) {
-                      return Container(
-                        margin: const EdgeInsets.symmetric(horizontal: 4.0),
-                        width: _currentPage == index ? 16.0 : 4.0,
-                        height: 4.0,
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(2),
-                          color: _currentPage == index
-                              ? Colors.white
-                              : Colors.grey,
-                        ),
-                      );
-                    },
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        Container(
-          width: 175,
-          height: 60,
-          decoration: const BoxDecoration(
-            borderRadius: BorderRadius.vertical(
-              top: Radius.zero,
-              bottom: Radius.circular(10),
-            ),
-            color: Colors.white,
-          ),
-          child: Padding(
-            padding:
-                const EdgeInsets.symmetric(horizontal: 10.0, vertical: 8.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const Text(
-                  '카페 랑데뷰',
-                  style: TextStyle(
-                    fontSize: 12,
-                    fontWeight: FontWeight.w400,
-                    color: Colors.black87,
-                  ),
-                ),
-                Row(
-                  children: [
-                    Container(
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(2),
-                        color: ThemeColor.basicGrey,
-                      ),
-                      child: const Padding(
-                        padding: EdgeInsets.all(2.0),
-                        child: Text(
-                          '#노을',
-                          style: TextStyle(
-                            fontSize: 12,
-                            color: Colors.black87,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
