+++ lib/src/presentation/course/bloc/main/main_event.dart
@@ -1,0 +1,5 @@
+sealed class MainEvent {}
+
+final class MainUnfocus extends MainEvent {}
+
+final class MainFocus extends MainEvent {}
