+++ lib/src/presentation/course/bloc/main/main_state.dart
