+++ lib/src/presentation/course/page/main_page.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:solo_play_application/src/presentation/course/bloc/main/main_bloc.dart';
+import 'package:solo_play_application/src/presentation/course/view/main_view.dart';
+
+class MainPage extends StatelessWidget {
+  const MainPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => MainBloc(),
+      child: const MainView(),
+    );
+  }
+}
