+++ test/app/view/app_view_bottom_nav_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:solo_play_application/src/app/presentation/page/app_page.dart';
+import 'package:solo_play_application/src/presentation/app/page/app_page.dart';
 
 void main() {
   group("App View Bottom Nav UI Test", () {
