+++ test/course/widget/course_widget_test.dart
@@ -1,18 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:solo_play_application/src/course/presentation/widget/course_widget.dart';
+import 'package:solo_play_application/src/presentation/course/widget/course_widget.dart';
 
 void main() {
-  late final widget;
+  late final MaterialApp widget;
 
   group("Course Widget Tests", () {
     setUpAll(() {
-      widget = MaterialApp(home: CourseWidget());
+      widget = const MaterialApp(home: CourseWidget());
     });
     testWidgets("Course Widget이 올바르게 화면에 표시된다.", (tester) async {
       await tester.pumpWidget(widget);
-      expect(find.byKey(Key("Course Widget Map")), findsOneWidget);
-      expect(find.byKey(Key("Course Widget Description")), findsOneWidget);
+      expect(find.byKey(const Key("Course Widget Map")), findsOneWidget);
+      expect(
+          find.byKey(const Key("Course Widget Description")), findsOneWidget);
     });
 
     testWidgets("Course Widget은 지역이 올바르게 표시된다.", (tester) async {
