+++ test/main/widget/hexagon_position_test.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:solo_play_application/src/course/presentation/widget/hexagon_grid.dart';
+import 'package:solo_play_application/src/presentation/course/widget/hexagon_grid.dart';
 
 void main() {
   group('Hexagon Position Unit Test', () {
