+++ test/user/my_profile_view_ui_test.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:solo_play_application/src/user/presentation/view/my_profile_view.dart';
+import 'package:solo_play_application/src/presentation/user/view/my_profile_view.dart';
 
 void main() {
-  late final widget;
+  late final MaterialApp widget;
 
   group("My Profile View UI Test", () {
     setUpAll(() {
-      widget = MaterialApp(
+      widget = const MaterialApp(
         home: MyProfileView(),
       );
     });
